Fixed IDE import autocomplete
