+++ services/apps/integration_data_worker/src/conf/index.ts
@@ -1,22 +1,17 @@
 import { IDatabaseConfig } from '@crowd/database'
 import { IRedisConfiguration } from '@crowd/redis'
 import { ISqsClientConfig } from '@crowd/sqs'
-import config = require('config')
+import config from 'config'
 
-export interface IWorkerConfig {
-  maxStreamRetries: number
+export interface IWorkerSettings {
+  maxDataRetries: number
 }
 
-export interface INangoConfig {
-  url: string
-  secretKey: string
-}
-
-let workerSettings: IWorkerConfig
-export const WORKER_SETTINGS = (): IWorkerConfig => {
+let workerSettings: IWorkerSettings
+export const WORKER_SETTINGS = (): IWorkerSettings => {
   if (workerSettings) return workerSettings
 
-  workerSettings = config.get<IWorkerConfig>('worker')
+  workerSettings = config.get<IWorkerSettings>('worker')
   return workerSettings
 }
 
@@ -42,14 +37,6 @@ export const REDIS_CONFIG = (): IRedisConfiguration => {
 
   redisConfig = config.get<IRedisConfiguration>('redis')
   return redisConfig
-}
-
-let nangoConfig: INangoConfig
-export const NANGO_CONFIG = (): INangoConfig => {
-  if (nangoConfig) return nangoConfig
-
-  nangoConfig = config.get<INangoConfig>('nango')
-  return nangoConfig
 }
 
 let platformConfig: unknown | null | undefined = null
