+++ services/apps/integration_data_worker/src/main.ts
@@ -1,5 +1,5 @@
 import { getServiceLogger } from '@crowd/logging'
-import { DB_CONFIG, REDIS_CONFIG, SQS_CONFIG } from './config'
+import { DB_CONFIG, REDIS_CONFIG, SQS_CONFIG } from './conf'
 import { getRedisClient } from '@crowd/redis'
 import { getDbConnection } from '@crowd/database'
 import { DataSinkWorkerEmitter, IntegrationStreamWorkerEmitter, getSqsClient } from '@crowd/sqs'
