+++ services/apps/integration_run_worker/src/main.ts
@@ -1,7 +1,7 @@
 import { getDbConnection } from '@crowd/database'
 import { getServiceLogger } from '@crowd/logging'
 import { IntegrationStreamWorkerEmitter, getSqsClient } from '@crowd/sqs'
-import { DB_CONFIG, REDIS_CONFIG, SQS_CONFIG } from './config'
+import { DB_CONFIG, REDIS_CONFIG, SQS_CONFIG } from './conf'
 import { WorkerQueueReceiver } from './queue'
 import { ApiPubSubEmitter, getRedisClient } from '@crowd/redis'
