+++ services/libs/common/src/index.ts
@@ -3,6 +3,6 @@ export * from './timing'
 export * from './utils'
 export * from './array'
 export * from './object'
-export * from './uuid'
+export * from './uuidUtils'
 export * from './validations'
 export * from './strings'
